requirements.txt
streamlit
pyodbc
pandas
requests
python-dotenv